+++ src/Module/OxideshopModules.php
@@ -1,51 +1,76 @@
 <?php
+
 /**
  * Copyright © OXID eSales AG. All rights reserved.
  * See LICENSE file for license details.
  */
+
+declare(strict_types=1);
 
 namespace OxidEsales\Codeception\Module;
 
 require_once __DIR__.'/../../../../oxid-esales/testing-library/base.php';
 
 use Codeception\Lib\ModuleContainer;
+use Codeception\Module as CodeceptionModule;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Setup\Exception\ModuleSetupException;
 use OxidEsales\Facts\Facts;
+use Symfony\Component\Console\Command\Command;
 
-/**
- * Class Oxideshop
- * @package OxidEsales\Codeception\Module
- */
-class OxideshopModules extends \Codeception\Module
+class OxideshopModules extends CodeceptionModule
 {
     /** @var string */
-    private $shopRootPath;
+    private $consoleRunner;
 
-    /**
-     * @var string
-     */
-    private $communityEditionRootPath;
-
-    public function __construct(ModuleContainer $moduleContainer, $config = null)
-    {
-        $this->shopRootPath = (new Facts())->getShopRootPath();
-        $this->communityEditionRootPath = (new Facts())->getCommunityEditionRootPath();
+    public function __construct(
+        ModuleContainer $moduleContainer,
+        $config = null
+    ) {
+        $shopRoot = (new Facts())->getCommunityEditionRootPath();
+        $this->consoleRunner = "$shopRoot/bin/oe-console";
 
         parent::__construct($moduleContainer, $config);
     }
 
-    /**
-     * Reset context and activate modules before test
-     */
-    public function _beforeSuite($settings = [])
+    public function _beforeSuite($settings = []): void
     {
         $this->activateModules();
+    }
+
+    /** @param string $modulePath */
+    public function installModule(string $modulePath): void
+    {
+        exec("$this->consoleRunner oe:module:install $modulePath", $output, $return);
+        if ($return !== Command::SUCCESS) {
+            throw new \RuntimeException("Module '$modulePath' installation failed: $output");
+        }
+    }
+
+    /** @param string $moduleId */
+    public function uninstallModule(string $moduleId): void
+    {
+        exec("$this->consoleRunner oe:module:uninstall $moduleId", $output, $return);
+        if ($return !== Command::SUCCESS) {
+            throw new \RuntimeException("Module '$moduleId' uninstallation failed: $output");
+        }
+    }
+
+    /** @param string $moduleId */
+    public function activateModule(string $moduleId): void
+    {
+        exec("$this->consoleRunner oe:module:activate $moduleId");
+    }
+
+    /** @param string $moduleId */
+    public function deactivateModule(string $moduleId): void
+    {
+        exec("$this->consoleRunner oe:module:deactivate $moduleId");
     }
 
     /**
      * Activates modules
      */
-    private function activateModules()
+    private function activateModules(): void
     {
         $testConfig = new \OxidEsales\TestingLibrary\TestConfig();
         $modulesToActivate = $testConfig->getModulesToActivate();
@@ -60,45 +85,5 @@ class OxideshopModules extends \Codeception\Module
                 // we can ignore this
             }
         }
-    }
-
-    public function getShopModulePath(string $modulePath): string
-    {
-        return $this->shopRootPath . '/source/modules' . substr($modulePath, strrpos($modulePath, '/'));
-    }
-
-    public function installModule($modulePath)
-    {
-        //first Copy
-        exec('cp ' . $modulePath . ' ' . $this->shopRootPath . '/source/modules/ -R');
-        //now activate
-        exec(
-            $this->communityEditionRootPath .
-            '/bin/oe-console oe:module:install-configuration ' .
-            $this->getShopModulePath($modulePath)
-        );
-    }
-
-    public function uninstallModule($modulePath, $moduleId)
-    {
-        exec(
-            $this->communityEditionRootPath .
-            '/bin/oe-console oe:module:uninstall-configuration ' .
-            $moduleId
-        );
-        $path = $this->getShopModulePath($modulePath);
-        if (file_exists($path) && is_dir($path)) {
-            exec('rm ' . $path . ' -R');
-        }
-    }
-
-    public function activateModule($moduleId)
-    {
-        exec($this->communityEditionRootPath . '/bin/oe-console oe:module:activate ' . $moduleId);
-    }
-
-    public function deactivateModule($moduleId)
-    {
-        exec($this->communityEditionRootPath . '/bin/oe-console oe:module:deactivate ' . $moduleId);
     }
 }